warning fix

+++ src/gui/desktop/model_view/view_helpers/view_data_set.hpp
@@ -233,7 +233,7 @@ class ViewDataSet final: ModelObserverInterface
             auto parentIt = find(parent);
             flat_iterator flat_parent(parentIt);
 
-            if (flat_parent.children_count() > to)
+            if (flat_parent.children_count() > static_cast<unsigned int>(to))
             {
                 for(int i = from; i <= to; i++)
                 {
